feat: структура проекта, Header, Logo, Navbar
