+++ src/components/Navbar.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export const Navbar = () => {
+  return (
+    <nav>
+      <a href="#">Карта</a>
+      <a href="#">Профиль</a>
+      <a href="#">Выйти</a>
+    </nav>
+  );
+};
