+++ src/pages/Map.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Header } from '../components/Header'; 
+
+export const Map = () => {
+  return (
+    <>
+      <Header />
+    </>
+  );
+};
